requirements.txt
pandas
altair
streamlit_pdf_viewer